requirements.txt
gradio
python-dotenv
google-generativeai

